Small adjustments to basic encoder file for the future
